When a builtin is being run in the parent process the SDTIN and STDOUT are being restored before looping to next commandline, also moved functions from one big file into a few smaller ones
