+++ sources/execution/execute_commandline.c
@@ -6,53 +6,11 @@
 /*   By: maheleni <maheleni@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/30 10:33:14 by maheleni          #+#    #+#             */
-/*   Updated: 2024/10/10 16:32:55 by maheleni         ###   ########.fr       */
+/*   Updated: 2024/10/11 10:33:52 by maheleni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
-
-int	execute_builtin(t_main *main, t_tokens token)
-{
-	char	*command;
-	int		cmd_len;
-
-	(void)main;
-	command = token.command[0];
-	cmd_len = ft_strlen(command);
-	if (ft_strncmp(command, "echo\0", cmd_len) == 0)
-	{
-		return (echo(main, token));
-	}
-	else if (ft_strncmp(command, "cd\0", cmd_len) == 0)
-	{
-		// return (cd());
-		printf("Executing cd\n");
-	}
-	else if (ft_strncmp(command, "pwd\0", cmd_len) == 0)
-	{
-		return (pwd(main, token));
-	}
-	else if (ft_strncmp(command, "export\0", cmd_len) == 0)
-	{
-		return (export(main, token));
-	}
-	else if (ft_strncmp(command, "unset\0", cmd_len) == 0)
-	{
-		return (unset(main, token));
-	}
-	else if (ft_strncmp(command, "env\0", cmd_len) == 0)
-	{
-		return (env(main, token));
-	}
-	else if (ft_strncmp(command, "exit\0", cmd_len) == 0)
-	{
-		printf("Executing exit\n");
-		//return(exit_command(main));
-		return (0);
-	}
-	return (0);
-}
 
 int	is_builtin(t_tokens token)
 {
@@ -83,133 +41,6 @@ int	is_builtin(t_tokens token)
 	return (0);
 }
 
-void	execute_command(t_main *main, t_tokens token)
-{
-	char	*path;
-	// char	**env;
-
-	path = get_path(main, token.command);
-	if (path == NULL)
-	{
-		dup2(STDERR_FILENO, STDOUT_FILENO);
-		if (errno == 127)
-			printf("Command not found: %s\n", token.command[0]);
-		else
-			perror("Error in getting path to command:");
-		exit(1);
-	}
-
-	//make sure that in- and outfiles are correct at this point and pipes closed
-
-	// env = transform_to_string_array(main->env_list);		//TODO
-	// if (execve(path, token.command, env) == -1)
-	// {
-	// 	free(path);
-	// 	//error
-	// }
-
-	//this is just for testing without env
-	if (execv(path, token.command) == -1)
-	{
-		//make sure that path is always allocated by execution
-		printf("Execv error on command %s\n", token.command[0]);
-		free(path);
-		//error
-	}
-}
-
-void	dup2_outfile(t_tokens token)
-{
-	int				outfile;
-	t_redirect_node	*node;
-
-	node = token.outfiles;
-	while (node != NULL)
-	{
-		if (node->type == OUTFILE)
-			outfile = open(node->name, O_WRONLY | O_TRUNC | O_CREAT, 0644);
-		else if (node->type == APPEND)
-			outfile = open(node->name, O_WRONLY | O_APPEND | O_CREAT, 0644);		//CHECK THAT IS CORRECT
-		if (outfile == -1)
-		{
-			perror("In dup2_outfile");
-			exit(1);
-		}
-		node = node->next;
-	}
-	if (dup2(outfile, STDOUT_FILENO) == -1)
-	{
-		perror("dup2 in dup2_outfile");
-		//TODO is this correct?
-		exit(1);
-	}
-	close(outfile);
-}
-
-void	handle_outfile(t_tokens token, int* pipe_right)
-{
-	if (pipe_right != NULL && token.outfiles == NULL)
-	{
-		if (dup2(pipe_right[1], STDOUT_FILENO) == -1)
-		{
-			perror("dup2 in handle_outfile");
-			//TODO is this correct?
-			exit(1);
-		}
-		close(pipe_right[0]);
-		close(pipe_right[1]);
-	}
-	if (token.outfiles != NULL)
-		dup2_outfile(token);
-}
-
-void	dup2_infile(t_tokens token)
-{
-	int	infile;
-	t_redirect_node	*node;
-
-	node = token.infiles;
-	while(node != NULL)
-	{
-		if (node != token.infiles)
-			close(infile);
-		infile = open(node->name, O_RDONLY);
-		if (infile == -1)
-		{
-			//TODO handle error
-			perror("In dup2_infile");
-			exit(1);
-		}
-		node = node->next;
-	}
-	if (dup2(infile, STDIN_FILENO) == -1)
-	{
-		perror("dup2 in dup2_infile");
-		//TODO is this correct?
-		exit(1);
-	}
-	close(infile);
-}
-
-void	handle_infile(t_tokens token, int* pipe_left)
-{
-	if (pipe_left != NULL && token.infiles == NULL)
-	{
-		if (dup2(pipe_left[0], STDIN_FILENO) == -1)
-		{
-			perror("dup2 in handle_infile");
-			//TODO is this correct?
-			exit(1);
-		}
-		close(pipe_left[0]);
-		close(pipe_left[1]);
-	}
-	if (token.infiles != NULL)
-	{
-		dup2_infile(token);
-	}
-}
-
 //pipe[0]: read, pipe[1]: write
 void	close_pipes_in_parent(int i, int num_of_pipes, int *pipe_left, int *pipe_right)
 {
@@ -221,30 +52,6 @@ void	close_pipes_in_parent(int i, int num_of_pipes, int *pipe_left, int *pipe_ri
 	{
 		close(pipe_left[0]);
 		close(pipe_right[1]);
-	}
-}
-
-void	handle_infile_and_outfile(int i, int num_of_pipes, int pipe_array[2][2], t_tokens token)
-{
-	if (i == 0 && num_of_pipes == 0)
-	{
-		handle_infile(token, NULL);
-		handle_outfile(token, NULL);
-	}
-	else if (i > 0 && num_of_pipes == 0)
-	{
-		handle_infile(token, pipe_array[0]);
-		handle_outfile(token, NULL);
-	}
-	else if (i == 0)
-	{
-		handle_infile(token, NULL);
-		handle_outfile(token, pipe_array[1]);
-	}
-	else
-	{
-		handle_infile(token, pipe_array[0]);
-		handle_outfile(token, pipe_array[1]);
 	}
 }
 
@@ -285,33 +92,10 @@ int	create_fork(void)
 	return (pid);
 }
 
-void	execute_child_process(t_main *main, t_tokens token)
-{
-	int	status;
-
-	if (is_builtin(token))
-	{
-		status = execute_builtin(main, token);
-		exit(status);
-	}
-	execute_command(main, token);
-}
-
 void	reassign_pipe_right_to_left(int pipe_array[2][2])
 {
 	pipe_array[0][0] = pipe_array[1][0];
 	pipe_array[0][1] = pipe_array[1][1];
-}
-
-void	execute_builtin_in_parent(int i, t_tokens *tokens, t_main *main)
-{
-	int	outfile_fd;
-
-	outfile_fd = -1;
-	handle_infile(tokens[i], NULL);
-	handle_outfile(tokens[i], NULL);
-	execute_builtin(main, tokens[i]);
-	//work here
 }
 
 int	is_builtin_not_part_of_pipeline(t_tokens *tokens, int num_of_pipes, int i)
@@ -335,7 +119,7 @@ int	execute_commandline(t_main *main, t_tokens *tokens)
 	while (num_of_pipes >= 0)
 	{
 		if (is_builtin_not_part_of_pipeline(tokens, num_of_pipes, i))
-			execute_builtin_in_parent(i, tokens, main);
+			execute_builtin_in_parent(main, tokens[i]);
 		else
 		{
 			if (i > 0)
